# ui/src/lib/api/openapi.json
{"openapi":"3.1.0","info":{"title":"Workflow Execution Service","version":"0.1.0"},"paths":{"/event":{"post":{"summary":" Handle Event Post","operationId":"_handle_event_post_event_post","requestBody":{"content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/HttpWorkflowUpdateEvent"},{"$ref":"#/components/schemas/HttpRecordingStartedEvent"},{"$ref":"#/components/schemas/HttpRecordingStoppedEvent"}],"title":"Event Data"}}},"required":true},"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/workflows":{"get":{"summary":"List Workflows","operationId":"list_workflows_api_workflows_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkflowListResponse"}}}}}}},"/api/workflows/{name}":{"get":{"summary":"Get Workflow","operationId":"get_workflow_api_workflows__name__get","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string","title":"Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"string","title":"Response Get Workflow Api Workflows  Name  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Workflow","operationId":"delete_workflow_api_workflows__name__delete","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string","title":"Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkflowResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/workflows/update":{"post":{"summary":"Update Workflow","operationId":"update_workflow_api_workflows_update_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkflowUpdateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkflowResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/workflows/update-metadata":{"post":{"summary":"Update Workflow Metadata","operationId":"update_workflow_metadata_api_workflows_update_metadata_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkflowMetadataUpdateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkflowResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/workflows/delete-step":{"post":{"summary":"Delete Step","operationId":"delete_step_api_workflows_delete_step_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkflowDeleteStepRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkflowResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/workflows/execute":{"post":{"summary":"Execute Workflow","operationId":"execute_workflow_api_workflows_execute_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkflowExecuteRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkflowExecuteResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/workflows/logs/{task_id}":{"get":{"summary":"Get Logs","operationId":"get_logs_api_workflows_logs__task_id__get","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}},{"name":"position","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Position"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkflowLogsResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/workflows/tasks/{task_id}/status":{"get":{"summary":"Get Task Status","operationId":"get_task_status_api_workflows_tasks__task_id__status_get","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkflowStatusResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/workflows/tasks/{task_id}/cancel":{"post":{"summary":"Cancel Workflow","operationId":"cancel_workflow_api_workflows_tasks__task_id__cancel_post","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkflowCancelResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/workflows/add":{"post":{"summary":"Add Workflow","operationId":"add_workflow_api_workflows_add_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkflowAddRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkflowResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/workflows/record":{"post":{"summary":"Record Workflow","operationId":"record_workflow_api_workflows_record_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkflowRecordResponse"}}}}}}},"/api/workflows/cancel-recording":{"post":{"summary":"Cancel Recording","operationId":"cancel_recording_api_workflows_cancel_recording_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkflowRecordResponse"}}}}}}},"/api/workflows/build-from-recording":{"post":{"summary":"Build Workflow","operationId":"build_workflow_api_workflows_build_from_recording_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkflowBuildRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkflowBuildResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AgentTaskWorkflowStep":{"properties":{"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","description":"Optional description/comment about the step's purpose."},"output":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Output","description":"Context key to store step output under."},"type":{"type":"string","const":"agent","title":"Type"},"task":{"type":"string","title":"Task","description":"The objective or task description for the agent."},"max_steps":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Max Steps","description":"Maximum number of iterations for the agent (default handled in code)."}},"additionalProperties":true,"type":"object","required":["type","task"],"title":"AgentTaskWorkflowStep"},"ClickStep":{"properties":{"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","description":"Optional description/comment about the step's purpose."},"output":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Output","description":"Context key to store step output under."},"timestamp":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Timestamp","description":"Timestamp from recording (informational)."},"tabId":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Tabid","description":"Browser tab ID from recording (informational)."},"type":{"type":"string","const":"click","title":"Type"},"cssSelector":{"type":"string","title":"Cssselector","description":"CSS selector for the target element."},"xpath":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Xpath","description":"XPath selector (often informational)."},"elementTag":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Elementtag","description":"HTML tag (informational)."},"elementText":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Elementtext","description":"Element text (informational)."}},"additionalProperties":true,"type":"object","required":["type","cssSelector"],"title":"ClickStep","description":"Clicks an element using 'click' (maps to workflow controller's click)."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HttpRecordingStartedEvent":{"properties":{"timestamp":{"type":"integer","title":"Timestamp"},"type":{"type":"string","const":"RECORDING_STARTED","title":"Type","default":"RECORDING_STARTED"},"payload":{"$ref":"#/components/schemas/RecordingStatusPayload"}},"type":"object","required":["timestamp","payload"],"title":"HttpRecordingStartedEvent"},"HttpRecordingStoppedEvent":{"properties":{"timestamp":{"type":"integer","title":"Timestamp"},"type":{"type":"string","const":"RECORDING_STOPPED","title":"Type","default":"RECORDING_STOPPED"},"payload":{"$ref":"#/components/schemas/RecordingStatusPayload"}},"type":"object","required":["timestamp","payload"],"title":"HttpRecordingStoppedEvent"},"HttpWorkflowUpdateEvent":{"properties":{"timestamp":{"type":"integer","title":"Timestamp"},"type":{"type":"string","const":"WORKFLOW_UPDATE","title":"Type","default":"WORKFLOW_UPDATE"},"payload":{"$ref":"#/components/schemas/WorkflowDefinitionSchema"}},"type":"object","required":["timestamp","payload"],"title":"HttpWorkflowUpdateEvent"},"InputStep":{"properties":{"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","description":"Optional description/comment about the step's purpose."},"output":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Output","description":"Context key to store step output under."},"timestamp":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Timestamp","description":"Timestamp from recording (informational)."},"tabId":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Tabid","description":"Browser tab ID from recording (informational)."},"type":{"type":"string","const":"input","title":"Type"},"cssSelector":{"type":"string","title":"Cssselector","description":"CSS selector for the target input element."},"value":{"type":"string","title":"Value","description":"Value to input. Can use {context_var}."},"xpath":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Xpath","description":"XPath selector (informational)."},"elementTag":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Elementtag","description":"HTML tag (informational)."}},"additionalProperties":true,"type":"object","required":["type","cssSelector","value"],"title":"InputStep","description":"Inputs text using 'input' (maps to workflow controller's input)."},"KeyPressStep":{"properties":{"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","description":"Optional description/comment about the step's purpose."},"output":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Output","description":"Context key to store step output under."},"timestamp":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Timestamp","description":"Timestamp from recording (informational)."},"tabId":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Tabid","description":"Browser tab ID from recording (informational)."},"type":{"type":"string","const":"key_press","title":"Type"},"cssSelector":{"type":"string","title":"Cssselector","description":"CSS selector for the target element."},"key":{"type":"string","title":"Key","description":"The key to press (e.g., 'Tab', 'Enter')."},"xpath":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Xpath","description":"XPath selector (informational)."},"elementTag":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Elementtag","description":"HTML tag (informational)."}},"additionalProperties":true,"type":"object","required":["type","cssSelector","key"],"title":"KeyPressStep","description":"Presses a key using 'key_press' (maps to workflow controller's key_press)."},"NavigationStep":{"properties":{"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","description":"Optional description/comment about the step's purpose."},"output":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Output","description":"Context key to store step output under."},"timestamp":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Timestamp","description":"Timestamp from recording (informational)."},"tabId":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Tabid","description":"Browser tab ID from recording (informational)."},"type":{"type":"string","const":"navigation","title":"Type"},"url":{"type":"string","title":"Url","description":"Target URL to navigate to. Can use {context_var}."}},"additionalProperties":true,"type":"object","required":["type","url"],"title":"NavigationStep","description":"Navigates using the 'navigation' action (likely maps to go_to_url)."},"PageExtractionStep":{"properties":{"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","description":"Optional description/comment about the step's purpose."},"output":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Output","description":"Context key to store step output under."},"timestamp":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Timestamp","description":"Timestamp from recording (informational)."},"tabId":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Tabid","description":"Browser tab ID from recording (informational)."},"type":{"type":"string","const":"extract_page_content","title":"Type"},"goal":{"type":"string","title":"Goal","description":"The goal of the page extraction."}},"additionalProperties":true,"type":"object","required":["type","goal"],"title":"PageExtractionStep","description":"Extracts text from the page using 'page_extraction' (maps to workflow controller's page_extraction)."},"RecordingStatusPayload":{"properties":{"message":{"type":"string","title":"Message"}},"type":"object","required":["message"],"title":"RecordingStatusPayload"},"ScrollStep":{"properties":{"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","description":"Optional description/comment about the step's purpose."},"output":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Output","description":"Context key to store step output under."},"timestamp":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Timestamp","description":"Timestamp from recording (informational)."},"tabId":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Tabid","description":"Browser tab ID from recording (informational)."},"type":{"type":"string","const":"scroll","title":"Type"},"scrollX":{"type":"integer","title":"Scrollx","description":"Horizontal scroll pixels."},"scrollY":{"type":"integer","title":"Scrolly","description":"Vertical scroll pixels."}},"additionalProperties":true,"type":"object","required":["type","scrollX","scrollY"],"title":"ScrollStep","description":"Scrolls the page using 'scroll' (maps to workflow controller's scroll)."},"SelectChangeStep":{"properties":{"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","description":"Optional description/comment about the step's purpose."},"output":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Output","description":"Context key to store step output under."},"timestamp":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Timestamp","description":"Timestamp from recording (informational)."},"tabId":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Tabid","description":"Browser tab ID from recording (informational)."},"type":{"type":"string","const":"select_change","title":"Type"},"cssSelector":{"type":"string","title":"Cssselector","description":"CSS selector for the target select element."},"selectedText":{"type":"string","title":"Selectedtext","description":"Visible text of the option to select. Can use {context_var}."},"xpath":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Xpath","description":"XPath selector (informational)."},"elementTag":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Elementtag","description":"HTML tag (informational)."}},"additionalProperties":true,"type":"object","required":["type","cssSelector","selectedText"],"title":"SelectChangeStep","description":"Selects a dropdown option using 'select_change' (maps to workflow controller's select_change)."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"WorkflowAddRequest":{"properties":{"name":{"type":"string","title":"Name"},"content":{"type":"string","title":"Content"}},"type":"object","required":["name","content"],"title":"WorkflowAddRequest"},"WorkflowBuildRequest":{"properties":{"workflow":{"$ref":"#/components/schemas/WorkflowDefinitionSchema"},"prompt":{"type":"string","title":"Prompt"},"name":{"type":"string","title":"Name"}},"type":"object","required":["workflow","prompt","name"],"title":"WorkflowBuildRequest"},"WorkflowBuildResponse":{"properties":{"success":{"type":"boolean","title":"Success"},"message":{"type":"string","title":"Message"},"error":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error"}},"type":"object","required":["success","message"],"title":"WorkflowBuildResponse"},"WorkflowCancelResponse":{"properties":{"success":{"type":"boolean","title":"Success"},"message":{"type":"string","title":"Message"}},"type":"object","required":["success","message"],"title":"WorkflowCancelResponse"},"WorkflowDefinitionSchema":{"properties":{"workflow_analysis":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Workflow Analysis","description":"A chain of thought reasoning analysis of the original workflow recording."},"name":{"type":"string","title":"Name","description":"The name of the workflow."},"description":{"type":"string","title":"Description","description":"A human-readable description of the workflow."},"version":{"type":"string","title":"Version","description":"The version identifier for this workflow definition."},"steps":{"items":{"anyOf":[{"$ref":"#/components/schemas/NavigationStep"},{"$ref":"#/components/schemas/ClickStep"},{"$ref":"#/components/schemas/InputStep"},{"$ref":"#/components/schemas/SelectChangeStep"},{"$ref":"#/components/schemas/KeyPressStep"},{"$ref":"#/components/schemas/ScrollStep"},{"$ref":"#/components/schemas/PageExtractionStep"},{"$ref":"#/components/schemas/AgentTaskWorkflowStep"}]},"type":"array","minItems":1,"title":"Steps","description":"An ordered list of steps (actions or agent tasks) to be executed."},"input_schema":{"items":{"$ref":"#/components/schemas/WorkflowInputSchemaDefinition"},"type":"array","title":"Input Schema","description":"List of input schema definitions."}},"type":"object","required":["name","description","version","steps","input_schema"],"title":"WorkflowDefinitionSchema","description":"Pydantic model representing the structure of the workflow JSON file."},"WorkflowDeleteStepRequest":{"properties":{"workflowName":{"type":"string","title":"Workflowname"},"stepIndex":{"type":"integer","title":"Stepindex"}},"type":"object","required":["workflowName","stepIndex"],"title":"WorkflowDeleteStepRequest"},"WorkflowExecuteRequest":{"properties":{"name":{"type":"string","title":"Name"},"inputs":{"type":"object","title":"Inputs"}},"type":"object","required":["name","inputs"],"title":"WorkflowExecuteRequest"},"WorkflowExecuteResponse":{"properties":{"success":{"type":"boolean","title":"Success"},"task_id":{"type":"string","title":"Task Id"},"workflow":{"type":"string","title":"Workflow"},"log_position":{"type":"integer","title":"Log Position"},"message":{"type":"string","title":"Message"}},"type":"object","required":["success","task_id","workflow","log_position","message"],"title":"WorkflowExecuteResponse"},"WorkflowInputSchemaDefinition":{"properties":{"name":{"type":"string","title":"Name","description":"The name of the property. This will be used as the key in the input schema."},"type":{"type":"string","enum":["string","number","bool"],"title":"Type"},"required":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Required","description":"None if the property is optional, True if the property is required."}},"type":"object","required":["name","type"],"title":"WorkflowInputSchemaDefinition"},"WorkflowListResponse":{"properties":{"workflows":{"items":{"type":"string"},"type":"array","title":"Workflows"}},"type":"object","required":["workflows"],"title":"WorkflowListResponse"},"WorkflowLogsResponse":{"properties":{"logs":{"items":{"type":"string"},"type":"array","title":"Logs"},"position":{"type":"integer","title":"Position"},"log_position":{"type":"integer","title":"Log Position"},"status":{"type":"string","title":"Status"},"result":{"anyOf":[{"items":{"type":"object"},"type":"array"},{"type":"null"}],"title":"Result"},"error":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error"}},"type":"object","required":["logs","position","log_position","status"],"title":"WorkflowLogsResponse"},"WorkflowMetadataUpdateRequest":{"properties":{"name":{"type":"string","title":"Name"},"metadata":{"type":"object","title":"Metadata"}},"type":"object","required":["name","metadata"],"title":"WorkflowMetadataUpdateRequest"},"WorkflowRecordResponse":{"properties":{"success":{"type":"boolean","title":"Success"},"workflow":{"anyOf":[{"$ref":"#/components/schemas/WorkflowDefinitionSchema"},{"type":"null"}]},"error":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error"}},"type":"object","required":["success"],"title":"WorkflowRecordResponse"},"WorkflowResponse":{"properties":{"success":{"type":"boolean","title":"Success"},"error":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error"}},"type":"object","required":["success"],"title":"WorkflowResponse"},"WorkflowStatusResponse":{"properties":{"task_id":{"type":"string","title":"Task Id"},"status":{"type":"string","title":"Status"},"workflow":{"type":"string","title":"Workflow"},"result":{"anyOf":[{"items":{"type":"object"},"type":"array"},{"type":"null"}],"title":"Result"},"error":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error"}},"type":"object","required":["task_id","status","workflow"],"title":"WorkflowStatusResponse"},"WorkflowUpdateRequest":{"properties":{"filename":{"type":"string","title":"Filename"},"nodeId":{"type":"integer","title":"Nodeid"},"stepData":{"type":"object","title":"Stepdata"}},"type":"object","required":["filename","nodeId","stepData"],"title":"WorkflowUpdateRequest"}}}}